tried some stuff
still cant figure out why the ragdoll wont fly. had a moment where it
was screwing up and sent it flying the wrong way but when i fixed that
it stopped flying all together. also cant seem to find the rigidbodys on
ragdoll could be im on the free version